visualisation
création de la visualisation des images mnist et mise en place de la fonction modèle pour train les modèle
